requirements.txt
asgiref==3.5.2
Django==4.1
django-mssql-backend==2.8.1
et-xmlfile==1.1.0
gunicorn==20.1.0
mysql-connector-python==8.0.30
openpyxl==3.0.10
protobuf==3.20.1
psycopg2==2.9.3
pyodbc==4.0.34
sqlparse==0.4.2
uWSGI==2.0.20
whitenoise==6.2.0
